+++ src/main/java/insane96mcp/iguanatweaksreborn/mixin/MobEffectMixin.java
@@ -2,7 +2,7 @@ package insane96mcp.iguanatweaksreborn.mixin;
 
 import com.llamalad7.mixinextras.injector.ModifyExpressionValue;
 import insane96mcp.iguanatweaksreborn.module.combat.MiscStats;
-import insane96mcp.iguanatweaksreborn.module.misc.tweaks.Tweaks;
+import insane96mcp.iguanatweaksreborn.module.misc.PotionsAndEffects;
 import insane96mcp.insanelib.base.Feature;
 import net.minecraft.world.effect.MobEffect;
 import org.spongepowered.asm.mixin.Mixin;
@@ -13,19 +13,19 @@ public class MobEffectMixin {
 
 	@ModifyExpressionValue(method = "isDurationEffectTick", at = @At(value = "CONSTANT", args = "intValue=25"))
 	public int onPoisonTickDamage(int poisonFactor) {
-		return Tweaks.getPoisonDamageSpeed(poisonFactor);
+		return PotionsAndEffects.getPoisonDamageSpeed(poisonFactor);
 	}
 
 	@ModifyExpressionValue(method = "applyEffectTick", at = @At(value = "CONSTANT", args = "intValue=4"))
 	public int onHealFactorTick(int healFactor) {
-		if (Feature.isEnabled(MiscStats.class) && MiscStats.betterHealingPotion)
+		if (Feature.isEnabled(MiscStats.class) && PotionsAndEffects.betterHealingPotion)
 			return 6;
 		return healFactor;
 	}
 
 	@ModifyExpressionValue(method = "applyInstantenousEffect", at = @At(value = "CONSTANT", args = "intValue=4"))
 	public int onHealFactorInstantaneous(int healFactor) {
-		if (Feature.isEnabled(MiscStats.class) && MiscStats.betterHealingPotion)
+		if (Feature.isEnabled(MiscStats.class) && PotionsAndEffects.betterHealingPotion)
 			return 6;
 		return healFactor;
 	}