+++ changelog.md
@@ -5,6 +5,9 @@
 * Blinker
   * New item that can be used to teleport straight forward where looking
   * Consumes durability and can be refilled with ender pearls
+* New Potions and Effects feature
+  * Added Streamline Splash Potions: when enabled will make splash potions always be 75% effective compared to drinkable ones but will also apply the full effect duration when thrown
+  * Moved all the relative config options here
 
 ### Changes
 * Split Stone tools disabling feature into its own